+++ src/simd/x86/sse_ops.c
@@ -483,7 +483,10 @@ void carquet_sse_byte_stream_split_decode_double(
  * Apply prefix sum (cumulative sum) to int32 array using SSE.
  */
 void carquet_sse_prefix_sum_i32(int32_t* values, int64_t count, int32_t initial) {
-    int32_t sum = initial;
+    /* Use unsigned arithmetic to avoid signed overflow UB.
+     * Delta encoding relies on modular arithmetic — _mm_add_epi32 is
+     * already modular, so only the scalar accumulator needs fixing. */
+    uint32_t sum = (uint32_t)initial;
     int64_t i = 0;
 
     /* SSE prefix sum for 4 elements at a time */
@@ -501,18 +504,18 @@ void carquet_sse_prefix_sum_i32(int32_t* values, int64_t count, int32_t initial)
         v = _mm_add_epi32(v, shifted2);
 
         /* Add running sum */
-        __m128i sums = _mm_set1_epi32(sum);
+        __m128i sums = _mm_set1_epi32((int32_t)sum);
         v = _mm_add_epi32(v, sums);
         _mm_storeu_si128((__m128i*)(values + i), v);
 
         /* Update running sum to last element */
-        sum = _mm_extract_epi32(v, 3);
+        sum = (uint32_t)_mm_extract_epi32(v, 3);
     }
 
     /* Handle remaining values */
     for (; i < count; i++) {
-        sum += values[i];
-        values[i] = sum;
+        sum += (uint32_t)values[i];
+        values[i] = (int32_t)sum;
     }
 }
 
@@ -520,7 +523,8 @@ void carquet_sse_prefix_sum_i32(int32_t* values, int64_t count, int32_t initial)
  * Apply prefix sum to int64 array using SSE.
  */
 void carquet_sse_prefix_sum_i64(int64_t* values, int64_t count, int64_t initial) {
-    int64_t sum = initial;
+    /* Use unsigned arithmetic to avoid signed overflow UB. */
+    uint64_t sum = (uint64_t)initial;
     int64_t i = 0;
 
     /* SSE prefix sum for 2 elements at a time */
@@ -532,18 +536,18 @@ void carquet_sse_prefix_sum_i64(int64_t* values, int64_t count, int64_t initial)
         v = _mm_add_epi64(v, shifted);
 
         /* Add running sum */
-        __m128i sums = _mm_set1_epi64x(sum);
+        __m128i sums = _mm_set1_epi64x((int64_t)sum);
         v = _mm_add_epi64(v, sums);
         _mm_storeu_si128((__m128i*)(values + i), v);
 
         /* Update running sum without spilling the whole vector */
-        sum = _mm_extract_epi64(v, 1);
+        sum = (uint64_t)_mm_extract_epi64(v, 1);
     }
 
     /* Handle remaining values */
     for (; i < count; i++) {
-        sum += values[i];
-        values[i] = sum;
+        sum += (uint64_t)values[i];
+        values[i] = (int64_t)sum;
     }
 }
